Lower Assign to store [assign] & Remove MandatoryOptUtils
update debug_type
